Fixing error relate to how the lvgl handle render while update through ble
