+++ main/inc/screen/home.h
@@ -4,14 +4,6 @@
 #include "lvgl.h"
 
 void home_screen(lv_obj_t *scr);
-
-#endif //__HOME_H__
-#ifndef __HOME_H__
-#define __HOME_H__
-
-#include "lvgl.h"
-
-void home_screen(lv_obj_t *scr);
 void home_update_steps(uint32_t steps);
 
 #endif //__HOME_H__
